requirements.txt
discord
youtube_dl
python-dotenv
asyncpraw
